+++ packages/lib/src/registry.ts
@@ -8,6 +8,7 @@ import CoreSettingSchema from './models/xernerx/core/Setting';
 import CoreInviteSchema from './models/xernerx/core/Invite';
 import CoreAppInviteSchema from './models/xernerx/core/AppInvite';
 import CoreRoleSchema from './models/xernerx/core/Role';
+import CoreAnnouncementSchema from './models/xernerx/core/Announcement';
 
 import DispatchApplicationSchema from './models/xernerx/dispatch/Application';
 import DispatchApplicationConfigSchema from './models/xernerx/dispatch/ApplicationConfig';
@@ -41,6 +42,7 @@ export const xernerxModels = {
 		Invite: { schema: CoreInviteSchema, modelName: 'Invite' },
 		AppInvite: { schema: CoreAppInviteSchema, modelName: 'AppInvite' },
 		Role: { schema: CoreRoleSchema, modelName: 'Role' },
+		Announcement: { schema: CoreAnnouncementSchema, modelName: 'Announcement' },
 	},
 	dispatch: {
 		Application: { schema: DispatchApplicationSchema, modelName: 'Application' },